pyproject.toml
[project]
name = "keyword-planner-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp[cli]",
    "google-ads",
    "python-dotenv>=1.1.0",
    "pandas>=2.2.3",
    "tabulate"
]

